Set format for expiration date
`Get-Date` string output is based on locale settings. Without any hints/formatting one would expect a long/long output and not short/long as we have seen in the e-mails.

This sets the formatting to `yyyy-MM-dd HH:mm`, which drops seconds as we do not need that kind of granularity.